testing/web-platform/tests/http/content_length.html

<!DOCTYPE html>

<title>Content-Length Test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
onload = function() {
  assert_equals(document.body.textContent, "PASS");
  done();
}
</script>
<body>PASS
but FAIL if this is in the body.
